writer: recenter API around runs, not event files
Users shouldn't have to think of event file names. This writer is now
more consistent with other TensorBoard event file writers (including the
official ones) in that it accepts a directory name and constructs the
event file name behind the scenes.

This is a breaking API change, but the old `new` method is still
available as `wrap`.